[Fix] Wow. It actually is working.
If you go to the GCMRegisterActivity and set the boolean at the top to true, it will open up your device to push notifications from my server. Right now, ANYONE in the world can go into that appspot and send a notification, so it is just for testing. Pretty cool though I think.

Not really sure where to go with it from here yet or how to get it there, but this is a start.